requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
openai==1.15.0
python-dotenv==1.0.1
tiktoken==0.6.0
